Navigator view

See merge request psege/rose!59

+++ src/main/java/edu/kit/rose/controller/roadsystem/RoseRoadSystemController.java
@@ -44,7 +44,7 @@ public class RoseRoadSystemController extends Controller
   }
 
   @Override
-  public void setZoomLevel(int zoomLevel) {
+  public void setZoomLevel(double zoomLevel) {
 
   }
